[Hyunjoong] solve 1002 quack
